Renamed file for JrPlatformWrongContentFormat.tsx
