+++ src/privateJrPlatform/Components/JrPlatformWrongContentFormat.tsx
